Added new application.js to core javascript files.
application.js sets up pathing for the webapp to avoid having to adjust
ajax calls throughout the application if the URL changes.  Instead of
referencing the url in a call as "/relative/path/action" you can now
just call it using "QT.WEBROOT + '/action'".

Updated scripts to reflect this change.