Changed volume chart to reflect the lowest unit of change instead of what the project was listed in.
Previously the volume chart was listed in whatever unit the main quota graph was labled in.  For instance, if the main graph was shown in GB, then the volume graph would show GB, even if it only changed by KB.  This was confusing since .000389 GB isn't as intuitive as 285 KB would be.