Updated the way that projects and their quota data is returned.
Previously we were using a hacky way to quickly get paged results of projects with the latest quota update.  Basically we would grab a page worth of projects, then get their latest quota updates from the last table separatly and join them together in a loop.  New method uses CakePHPs built in relations.  It still pulls a subquery, but the joining is done automatically instead of relying on the ordering to be correct in the loop.  A view was created for server stats for the reports page, there isn't a lot of time benefits, but its easier to pull from that rather than calculate everything manually.